+++ src/api/track/remove_liked_tracks.rs
@@ -11,12 +11,12 @@ use crate::{
 };
 
 pub struct RemoveLikedTracksOptions {
-    pub user_id: i32,
+    pub user_id: u64,
     pub track_ids: Vec<String>,
 }
 
 impl RemoveLikedTracksOptions {
-    pub fn new<S, I>(user_id: i32, track_ids: I) -> Self
+    pub fn new<S, I>(user_id: u64, track_ids: I) -> Self
     where
         S: Into<String>,
         I: IntoIterator<Item = S>,
