+++ src/model/playlist/library.rs
@@ -5,7 +5,7 @@ use crate::model::track::PartialTrack;
 #[derive(Debug, Eq, PartialEq, Clone, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Library {
-    pub uid: i32,
-    pub revision: i32,
+    pub uid: u64,
+    pub revision: u32,
     pub tracks: Vec<PartialTrack>,
 }
