+++ src/model/user/mod.rs
@@ -3,7 +3,7 @@ use serde::Deserialize;
 #[derive(Debug, Eq, PartialEq, Clone, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct User {
-    pub uid: i32,
+    pub uid: u64,
     pub login: String,
     pub name: Option<String>,
     pub display_name: Option<String>,
@@ -11,5 +11,5 @@ pub struct User {
     pub sex: Option<String>,
     pub verified: Option<bool>,
     #[serde(default)]
-    pub regions: Vec<i32>,
+    pub regions: Vec<u32>, 
 }
